AttachEntityToEntity on nonexistent doors
The thing is, the doors load a little late. If we attach too soon (too
far away), the doors haven't actually loaded yet and won't be attached
to. That leads to floating locks.
If we're too late (too close), then we risk not all locks coming with
the area.
"Fix" for now is to lower the area range and dynamically calculate the
center location.