I need to clean up this door/gate stuff
It needs to be properly unified into *just* doors, but using the
current gate code.
For that to happen, I have to rewrite the attachment logic for on-door
keypads? Maybe.

I at least have to nuke all occurrances of "door" and then replace all
"gate" with "door" and ...yeah, refactor.

Before that kind of refactor, I want me a rollback point.

Tag, you're it!